pyproject.toml
[project]
name = "finance-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=20.7",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
]
